innova_api/innova_api/requirements.txt
django
django-filter
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-cors-headers